Revert "Fix warning"
This reverts commit 658becf59a74767364ea25f360218e2b2aac41b8.

+++ tests/test-mongoc-rpc.c
@@ -57,7 +57,7 @@ assert_rpc_equal (const char   *filename,
    uint8_t *data;
    mongoc_iovec_t *iov;
    size_t length;
-   size_t off = 0;
+   off_t off = 0;
    int r;
    int i;
